Listen for URL changes in top frame and notify sidebar
Watch for URL changes made by the current web page using the History API
and notify the sidebar application.

This will enable the client to update the current set of annotations to
reflect the new URL and ensure that any new annotations created after
the URL change are associated with the current URL rather than the
original URL on which the page was loaded.