Always use injected `Document` object in `document.coffee`
The `Document` class for reading metadata from a DOM `Document` in
`document.coffee` (😭 at the naming) allows the DOM `Document` it reads
from to be specified as a constructor argument. This is mainly useful in
tests.

However a number of places in `document.coffee` used `$(<selector>)` to
create a jQuery collection in order to read properties and attributes of
eg. `<link>` and `<meta>` elements. Since jQuery doesn't have the
context of the injected DOM `Document`, this resulted in these values
always being read from the global `document` object.

This commit replaces such references with calls to
`@document.{querySelector, querySelectorAll}` so that we are always
reading from the injected DOM document.

This turned up a few bugs such as HTML elements not being appended to
`<head>` in tests and the wrong `rel` attribute being used for icon
links.